+++ pkg/driver/driverFactory.go
@@ -9,7 +9,7 @@ type IDriver interface {
 	SendCommandsSet(cmds []string, expectPattern string) (string, error)
 	FindDevicePrompt(regex string, pattern string) (string, error)
 	ReadUntil(pattern string) (string, error)
-	SetTimeout(timeout int)
+	SetTimeout(timeout uint8)
 }
 
 func NewDriver(Connection connections.Connection, Return string) IDriver {
